Separated special commands and commands into separate classes
